#320 [fix] URI 변경
