+++ module-api/src/main/java/com/mile/controller/moim/MoimControllerSwagger.java
@@ -370,17 +370,4 @@ public interface MoimControllerSwagger {
             @PathVariable("moimId") final String moimUrl
     );
 
-    @Operation(summary = "유저 글모임 리스트 조회")
-    @ApiResponses(
-            value = {
-                    @ApiResponse(responseCode = "200", description = "글모임 리스트 조회가 조회되었습니다.",
-                            content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
-                    @ApiResponse(responseCode = "401", description = "로그인 후 이용해주세요.",
-                            content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
-                    @ApiResponse(responseCode = "500", description = "서버 내부 오류입니다.",
-                            content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
-            }
-    )
-    ResponseEntity<SuccessResponse<MoimListOfUserResponse>> getMoimListOfUser();
-
 }
